+++ src/invoke-lambda/runtimes/node.js
@@ -1,18 +1,15 @@
 /* eslint-enable semi */
 /* eslint semi: [ 'error', 'always' ] */
-let {
-  projectSrc,
-  handlerFile,
-  handlerFunction,
-  shared,
-  views,
-} = JSON.parse(process.env.__ARC_CONFIG__);
-let context = JSON.parse(process.env.__ARC_CONTEXT__);
+let { __ARC_CONFIG__, __ARC_CONTEXT__ } = process.env;
+let { projectSrc, handlerFile, handlerFunction, shared, views } = JSON.parse(__ARC_CONFIG__);
+let context = JSON.parse(__ARC_CONTEXT__);
 let { join, sep } = require('path');
 let { existsSync: exists, readFileSync: read } = require('fs');
 let handler = './' + handlerFile;
 let fn = require(handler)[handlerFunction];
 let cwd = process.cwd();
+delete process.env.__ARC_CONFIG__;
+delete process.env.__ARC_CONTEXT__;
 
 function isPromise (obj) {
   return obj && typeof obj.then === 'function';