+++ src/invoke-lambda/runtimes/ruby.rb
@@ -2,11 +2,14 @@ require 'json'
 request = JSON.parse(STDIN.gets)
 
 config = JSON.parse(ENV['__ARC_CONFIG__'])
+context = JSON.parse(ENV['__ARC_CONTEXT__'])
+ENV.delete('__ARC_CONFIG__')
+ENV.delete('__ARC_CONTEXT__')
+
 index = './' + config['handlerFile']
 require index
 handlerFn = config['handlerFunction']
 
-context = JSON.parse(ENV['__ARC_CONTEXT__'])
 result = '__ARC__ '
 resultEnd = ' __ARC_END__'
 begin